- Renamed rosade -> diarcros - removed clojure depends for messages
