+++ diarcros_core/BUGS.org
@@ -1,0 +1,6 @@
+#+TITLE:     DIARCROS Bugs
+
+List bugs here
+
+* hanging during PR2 initialization
+* rosrun not found by exec
