+++ diarcros_core/README
@@ -1,0 +1,46 @@
+EAK: 3/25/2015
+Below documentation is out of date. Please refer to the hrilab DIARCROS wiki page.
+
+
+ROS/DIARC (v0.1)
+==============
+
+This package consits of ROS/DIARC bridge classes which allow other
+components to communicate with ROS.
+
+Class Generator
+---------------
+
+The class generator introspects into a running ROS system and allows
+you to generate an DIARC component which has bidirectional communication
+to a node. Currently it only looks for a ROS master in the default
+setup and does not support connecting to an arbitrary master (although
+exporting the`ROS_MASTER_URI` variable in your shell settings should
+allow it to overcome this problem.
+
+### Build ###
+
+To build, run `./gradlew installApp`.
+
+### Run  ###
+
+You can then run the ROS/DIARC class generator with:
+
+`./build/install/diarcros/bin/diarcros -m edu.tufts.hrilab.diarcros.core 'nodename'`
+
+or you can invoke the GUI
+
+`./build/install/diarcros/bin/diarcros -m edu.tufts.hrilab.diarcros.core -g`
+
+Generated Classes
+-----------------
+
+Generated class should be placed under `edu/tufts/hrilab/diarcros` and checked into
+source control. 
+
+TODO
+----
+
+- Support ROS services
+- Allow dyanmic_reconfig for nodes which support it
+- Allow modification of the ROS parameter tree
