+++ diarcros_core/proxy/launch/move_base_proxy.launch
