+++ diarcros_core/proxy/mainpage.dox
