+++ diarcros_core/proxy/manifest.xml
