--- rosade_core/BUGS.org
@@ -1,6 +1,0 @@
-#+TITLE:     ROSADE Bugs
-
-List bugs here
-
-* hanging during PR2 initialization
-* rosrun not found by exec
