--- rosade_core/README
@@ -1,46 +1,0 @@
-EAK: 3/25/2015
-Below documentation is out of date. Please refer to the hrilab ROSADE wiki page.
-
-
-ROS/ADE (v0.1)
-==============
-
-This package consits of ROS/ADE bridge classes which allow other
-components to communicate with ROS.
-
-Class Generator
----------------
-
-The class generator introspects into a running ROS system and allows
-you to generate an ADE component which has bidirectional communication
-to a node. Currently it only looks for a ROS master in the default
-setup and does not support connecting to an arbitrary master (although
-exporting the`ROS_MASTER_URI` variable in your shell settings should
-allow it to overcome this problem.
-
-### Build ###
-
-To build, run `./gradlew installApp`.
-
-### Run  ###
-
-You can then run the ROS/ADE class generator with:
-
-`./build/install/rosade/bin/rosade -m edu.tufts.hrilab.rosade.core 'nodename'`
-
-or you can invoke the GUI
-
-`./build/install/rosade/bin/rosade -m edu.tufts.hrilab.rosade.core -g`
-
-Generated Classes
------------------
-
-Generated class should be placed under `edu/tufts/hrilab/rosade` and checked into
-source control. 
-
-TODO
-----
-
-- Support ROS services
-- Allow dyanmic_reconfig for nodes which support it
-- Allow modification of the ROS parameter tree
