--- rosade_core/build.gradle
@@ -1,23 +1,0 @@
-plugins {
-  id 'maven-publish'
-}
-apply plugin: 'java-library'
-
-
-tasks.register('collectDeps', Copy) {
-  duplicatesStrategy 'exclude'
-  description 'Collect dependencies into a single location for simpler Java incovation'
-  group 'ROSADE'
-  dependsOn 'rosade_bootstrap:deps'
-  delete 'lib'
-  mkdir('lib')
-  from 'rosade_bootstrap/lib' into 'lib'
-  from 'rosade_bootstrap/build/libs' into 'lib'
-  from 'rosade_messages/build/libs' into 'lib'
-  from 'rosade/build/libs' into 'lib'
-}
-
-// Define a task to build, execute custom task, and publish
-tasks.register('collectAndPublish') {
-  dependsOn 'build', 'collectDeps', 'publishToMavenLocal'
-}
