+++ settings.gradle
@@ -15,7 +15,7 @@ include 'rosjava_bootstrap',
         'apache_xmlrpc_client',
         'apache_xmlrpc_server',
         'rosjava',
-        'rosjava_geometry',
+//        'rosjava_geometry',
         'actionlib_java',
-        'rosade_core:rosade_messages',
-        'rosade_core:rosade_bootstrap'
+        'diarcros_core:diarcros_messages',
+        'diarcros_core:diarcros_bootstrap'
